+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Fixed an issue where `.yarnproject` files couldn't be loaded if the C# project was trimmed.
+- Compiling a `CompilationJob` in type-check-only mode now includes the generated string table in its results.
+
 ### Removed
 
 ## [3.2.0] 2026-03-27